Support specifing truststore passwords
